Refactor VFS and Task struct
VFS used to snoop inside `taskList` and get the currently running
task by indexing it with `currentTaskIndex` (both in Scheduler).

Instead, refactor VFS so that there's a seperate instance of it
for each process.

Add `kernelVfs`, instantiated at VFS initialization, which serves
as the VFS instance the kernel can use.