Update scheduling algorithm
Before implementing process blocking, it was necessary to first
update the scheduling algorithm so it skips over blocked processes.
This commit implements this functionality.

While we're at it, refactor the scheduler so that instead of
reserving index 0 of the task index for the idle task, the idle
task is a global Task variable in the Scheduler translation unit.