Refactor Scheduler in preparation for multi-processing
Refactor Scheduler into a class, so that it can be instantiated.
The CPU struct added in this commit contains a pointer to the
Scheduler instance serving as the task scheduler for that core.
The static method Scheduler::GetScheduler returns the scheduler
instance responsible for the core on which it is called.

Scheduler.cpp contains the vector holding all of the CPU structs,
indexed by the core (processor) ID. It is assumed that BSP has
the core ID 0.

Refactor LAPIC into a class so that it, too, can be instantiated.
The Scheduler class contains a pointer to it. (LAPIC instances are
local to the each core)

Replace taskList with taskQueue, which is an instance of the class
Queue added in this commit. It is derived from Vector.

Delete SMP.cpp and SMP.h, which have been rendered obsolete by this
commit.