Refactor StringUtilities
Rather than returning a buffer in the kernel binary's .bss section,
refactor StringUtilities so that it takes an argument for the
string buffer that the caller can allocate in advance.

This makes it easier to write concurrent code, because the buffer
is no longer a global variable in StringUtilities at risk of being
modified concurrently by different cores.