Add multiprocessing support
Update Scheduler::StartCores so that it uses stivale2's SMP tool in
order to start non-BSP cores. These non-BSP cores integrate
seemlessly into the existing scheduling system.

Update CPU::GetCoreID so that it properly reads from the
IA32_TSC_AUX to get the ID of the core it is called on.