Update LAPIC timer to slightly improve precision
Rather than storing the LAPIC timer frequency in clocks per millisecond,
store it in Hz, which prevents the need to divide it by 1000 at
initialization.

Multiplying the frequency (stored in Hz) by the number of milliseconds
and then dividing the product by 1000 when calculating the reload count
is slightly more precise than dividing the frequency by 1000 at timer
initialization.