Implement TimerEntry expiration
Replace Scheduler::timerFireTimes with a TimerEntry vector.
Every time there is a timer interrupt, Scheduler::UpdateTimerEntries
is called from Scheduler::SwitchToNextTask to update all the timer
entries and deal with any that have expired.

When TimerEntry::unblockOnExpire is set to true, the task specified
by TimerEntry::pid will be unblocked when the TimerEntry expires.