Remove timer entry for constant Scheduler tick
Remove the timer entry in Scheduler::timerEntries used for the constant
tick of the Scheduler (100ms). Instead, configure the timer to fire
every 100ms if there are no closer (shorter time) timer entries when
updating the timer frequency in Scheduler::ConfigureTimerClosestExpiry.

This makes it easier to deal with context switches that occur outside of
the fixed constant Scheduler tick, because there is no need to remove
the constant Scheduler tick timer entry from the list and add a new one
to maintain constant ticking.