Fix Scheduler not updating time entries properly
Update Scheduler::UpdateTimerEntries so that it subtracts the time of
each timer entry by the actual time that has transpired since the last
Scheduler tick.

This fixes the bug where timer entries were not being updated properly
if Scheduler::UpdateTimerEntries is called outside of a normal timer
interrupt (e.g., exit system call), in which case
Scheduler::currentTimerTime does not correspond to the time passed since
the last Scheduler tick.

Add LAPIC::GetTimeRemainingMilliseconds which is used for calculate time
passed.