Add input buffering to Terminal
Add Terminal::InputCharacter which adds a character to the input buffer.
Update Terminal::Read so that it returns the input buffer if the
characters are already available, and if not, blocks the task requesting
the read until said characters are available.

Fix Terminal::Write so that it only prints the number of characters as
specified by the count argument.

Update Terminal so that it's a singleton, for easier access from a
keyboard interrupt.