Add incomplete stat/fstat system calls
Replace VFS::GetVnodeType with VFS::GetVnodeInfo, which returns the
newly added VFS::VnodeInfo struct containing the inode number, file size
and VFS::VnodeType of the vnode.

stat/fstat system calls call VFS::GetVnodeInfo, and the mlibc sysdeps
convert the returned VFS::VnodeInfo into the POSIX stat struct. The
VnodeInfo struct does not contain all of the fields of the POSIX stat
struct however, so this implementation of stat/fstat is incomplete.

Update the isatty system call implementation in the mlibc sysdeps so
that it uses the fstat system call.